+++ src/pytgpt/imager/__init__.py
@@ -1,6 +1,6 @@
 from .imager import Imager
-from .imager import session
+from .imager import Prodia
 
 __info__ = "Generate images using craiyon.com"
 
-__all__ = ["Imager", "session"]
+__all__ = ["Imager", "Prodia"]